the trigger worked! the bkgd script idea... not so much.
turns out the custom material for ARCameraBackground works best as an image/video cropped to screen res — like an iPhone camera feed. that's okay, i'll try disabling the bkgd script and changing the camera's clear flags to "skybox" instead. (btw, move the trigger to (0,0,0) of the portal and change the collider radius to 0.1 / 2.)